third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristic/gen-invalid-characteristic-name.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return getHealthThermometerService()
    .then(({service}) => {
      return assert_promise_rejects_with_message(
        service.getCharacteristic('wrong_name'),
        new DOMException(
          'Failed to execute \'getCharacteristic\' on ' +
          '\'BluetoothRemoteGATTService\': Invalid Characteristic name: ' +
          '\'wrong_name\'. ' +
          'It must be a valid UUID alias (e.g. 0x1234), ' +
          'UUID (lowercase hex characters e.g. ' +
          '\'00001234-0000-1000-8000-00805f9b34fb\'), ' +
          'or recognized standard name from ' +
          'https://www.bluetooth.com/specifications/gatt/characteristics' +
          ' e.g. \'aerobic_heart_rate_lower_limit\'.',
          'TypeError'),
        'Wrong Characteristic name passed.');
    });
}, 'Wrong Characteristic name. Reject with TypeError.');

</script>
